Removed internal IP address
Hi there! I was browsing through this really intriguing implementation against Kronos (thanks for open sourcing this code and showing what's possible against the Kronos API) and noticed that there may be an internal IP address listed in your configs.

I figured while I was here I'd remove in favor of a more generic address and add a comment. Feel free to reject / close if it's not going to work for you.

Thanks again and have a good one.